+++ src/main/java/br/com/api/offers/repository/IOfferService.java
@@ -1,12 +1,15 @@
-package br.com.api.offers.service;
+package br.com.api.offers.repository;
 
 import java.util.Date;
 import java.util.List;
+
+import org.springframework.stereotype.Repository;
 
 import br.com.api.offers.model.Offer;
 import br.com.api.offers.model.vo.OfferResponse;
 import br.com.api.offers.model.vo.OfferQty;
 
+@Repository
 public interface IOfferService {
 
     OfferResponse<OfferQty> getRecordCount(Date date);
